refactor: remove Tanstack Query hooks from /api to /hooks
