+++ src/api/index.ts
@@ -1,4 +1,3 @@
 export * from './axios';
 export * from './mocks';
 export * from './types';
-export * from './hooks';
