+++ src/hooks/useDeleteCart.ts
@@ -1,6 +1,6 @@
 import { QueryKey, useMutation, useQueryClient } from '@tanstack/react-query';
 import { CARTS_QUERY_KEY } from './useGetCarts';
-import { deleteCart } from '../axios';
+import { deleteCart } from '../api/axios';
 import type { CartResponse } from 'api/types';
 import type { ApiResponse } from 'api/types/ApiResponse';
