+++ src/hooks/useGetCarts.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getCarts } from '../axios/cartApi';
+import { getCarts } from '../api/axios/cartApi';
 
 export const CARTS_QUERY_KEY = 'carts';
