+++ src/hooks/useGetProducts.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getProducts } from '../axios/productApi';
+import { getProducts } from '../api/axios/productApi';
 
 const PRODUCTS_QUERY_KEY = 'products';
